Move version.updateSeriesVersions to be linux-only.
